show progress fix comment 1
