+++ app/src/main/AndroidManifest.xml
@@ -42,8 +42,8 @@
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".screen.basic_test.BasicTestActivity" />
 
-        <service android:name=".screen.splash.ImageService" />
-        <service android:name=".screen.splash.AudioService" />
+        <service android:name=".screen.service.ImageService" />
+        <service android:name=".screen.service.AudioService" />
     </application>
 
 </manifest>
