+++ app/src/main/java/com/framgia/toeic/screen/splash/SplashActivity.java
@@ -16,6 +16,8 @@ import com.framgia.toeic.data.repository.FileRepository;
 import com.framgia.toeic.data.source.remote.FileRemoteDatasource;
 import com.framgia.toeic.screen.base.BaseActivity;
 import com.framgia.toeic.screen.main.MainActivity;
+import com.framgia.toeic.screen.service.AudioService;
+import com.framgia.toeic.screen.service.ImageService;
 import com.google.firebase.FirebaseApp;
 
 import java.util.List;
